+++ src/routers/votesRouter.ts
@@ -99,8 +99,8 @@ votesRouter.post('/', async (req: Request, res: Response) => {
     await connection.beginTransaction();
 
     interface HangoutMemberDetails extends RowDataPacket {
-      current_stage: number,
       is_concluded: boolean,
+      current_stage: number,
       hangout_member_id: number,
       account_id: number | null,
       guest_id: number | null,
@@ -111,8 +111,8 @@ votesRouter.post('/', async (req: Request, res: Response) => {
 
     const [hangoutMemberRows] = await connection.execute<HangoutMemberDetails[]>(
       `SELECT
-        hangouts.current_stage,
         hangouts.is_concluded,
+        hangouts.current_stage,
         hangout_members.account_id,
         hangout_members.guest_id,
         EXISTS (SELECT 1 FROM suggestions WHERE suggestion_id = :suggestionId) AS suggestion_found,
@@ -125,7 +125,7 @@ votesRouter.post('/', async (req: Request, res: Response) => {
       WHERE
         hangouts.hangout_id = :hangoutId AND
         hangout_members.hangout_member_id = :hangoutMemberId
-        LIMIT 1;`,
+      LIMIT 1;`,
       {
         suggestionId: requestData.suggestionId,
         hangoutMemberId: requestData.hangoutMemberId,
@@ -153,12 +153,13 @@ votesRouter.post('/', async (req: Request, res: Response) => {
       return;
     };
 
-    if (hangoutMemberDetails.current_stage !== HANGOUT_VOTING_STAGE) {
-      await connection.rollback();
-      res.status(409).json({
-        message: hangoutMemberDetails.is_concluded ? 'Hangout already concluded' : `Hangout isn't in the voting stage.`
-      });
+    if (hangoutMemberDetails.is_concluded) {
+      res.status(403).json({ message: 'Hangout has already been concluded.', reason: 'hangoutConcluded' });
+      return;
+    };
 
+    if (hangoutMemberDetails.current_stage !== HANGOUT_VOTING_STAGE) {
+      res.status(403).json({ message: `Hangout hasn't reached the voting stage yet.`, reason: 'inAvailabilityStage' });
       return;
     };
 
@@ -349,8 +350,8 @@ votesRouter.delete('/', async (req: Request, res: Response) => {
     };
 
     interface HangoutMemberDetails extends RowDataPacket {
-      current_stage: number,
       is_concluded: boolean,
+      current_stage: number,
       hangout_member_id: number
       account_id: number | null,
       guest_id: number | null,
@@ -359,8 +360,8 @@ votesRouter.delete('/', async (req: Request, res: Response) => {
 
     const [hangoutMemberRows] = await dbPool.execute<HangoutMemberDetails[]>(
       `SELECT
-        hangouts.current_stage,
         hangouts.is_concluded,
+        hangouts.current_stage,
         hangout_members.account_id,
         hangout_members.guest_id,
         votes.vote_id
@@ -392,11 +393,13 @@ votesRouter.delete('/', async (req: Request, res: Response) => {
       return;
     };
 
-    if (hangoutMemberDetails.current_stage !== HANGOUT_VOTING_STAGE) {
-      res.status(409).json({
-        message: hangoutMemberDetails.is_concluded ? 'Hangout already concluded' : `Hangout isn't in the voting stage.`,
-      });
+    if (hangoutMemberDetails.is_concluded) {
+      res.status(403).json({ message: 'Hangout has already been concluded.', reason: 'hangoutConcluded' });
+      return;
+    };
 
+    if (hangoutMemberDetails.current_stage !== HANGOUT_VOTING_STAGE) {
+      res.status(403).json({ message: `Hangout hasn't reached the voting stage yet.`, reason: 'inAvailabilityStage' });
       return;
     };
 
@@ -507,8 +510,8 @@ votesRouter.delete('/clear', async (req: Request, res: Response) => {
     };
 
     interface HangoutMemberDetails extends RowDataPacket {
-      current_stage: number,
       is_concluded: boolean,
+      current_stage: number,
       account_id: number | null,
       guest_id: number | null,
       votes_id: number,
@@ -516,8 +519,8 @@ votesRouter.delete('/clear', async (req: Request, res: Response) => {
 
     const [hangoutMemberRows] = await dbPool.execute<HangoutMemberDetails[]>(
       `SELECT
-        hangouts.current_stage,
         hangouts.is_concluded,
+        hangouts.current_stage,
         hangout_members.account_id,
         hangout_members.guest_id,
         votes.vote_id
@@ -549,11 +552,13 @@ votesRouter.delete('/clear', async (req: Request, res: Response) => {
       return;
     };
 
-    if (hangoutMemberDetails.current_stage !== HANGOUT_VOTING_STAGE) {
-      res.status(409).json({
-        message: hangoutMemberDetails.is_concluded ? 'Hangout already concluded' : `Hangout isn't in the voting stage.`,
-      });
+    if (hangoutMemberDetails.is_concluded) {
+      res.status(403).json({ message: 'Hangout has already been concluded.', reason: 'hangoutConcluded' });
+      return;
+    };
 
+    if (hangoutMemberDetails.current_stage !== HANGOUT_VOTING_STAGE) {
+      res.status(403).json({ message: `Hangout hasn't reached the voting stage yet.`, reason: 'inAvailabilityStage' });
       return;
     };